fixing auth

+++ backend/api/user/router.py
@@ -1,10 +1,12 @@
 """ User management apis """
 
+from typing import Union
 from fastapi import APIRouter, Body, Header
 from pydantic import BaseModel
 from typing import Optional, Any
 from fastapi import Request, Depends
 from fastapi.security import HTTPBearer, HTTPAuthorizationCredentials
+from sqlalchemy import union
 from utils.jwt import decode_jwt, sign_jwt
 from utils.db import db, User
 from utils.error import Error
@@ -16,7 +18,7 @@ class JWTBearer(HTTPBearer):
     async def __call__(self, request: Request) -> Optional[str]:
         credentials: Optional[HTTPAuthorizationCredentials]  = await super(JWTBearer, self).__call__(request)
         if credentials and credentials.scheme == "Bearer":
-            return decode_jwt(credentials.credentials)["msisdn"]
+            return decode_jwt(credentials.credentials)
 
 router = APIRouter()
 
@@ -41,6 +43,11 @@ class UserRetrieve(BaseModel):
     email: Optional[str]
     profile_pic_url : Optional[str]
 
+
+def generate_unauth_error():
+    return {"status":"failed", "detail": "Not authenticated"}
+
+
 @router.post('/create', response_model=dict[str,Any])
 async def create_user(new_user: UserCreate) -> dict[str,Any]:
     """ Register a new user """
@@ -56,10 +63,18 @@ async def create_user(new_user: UserCreate) -> dict[str,Any]:
 
     return {"user_id": user.id} 
 
-@router.get('/profile', response_model=UserRetrieve, response_model_exclude_none=True)
-async def get_profile(msisdn = Depends(JWTBearer())) -> UserRetrieve:
+@router.get('/profile', response_model=Union[UserRetrieve, Any], response_model_exclude_none=True)
+async def get_profile(payload = Depends(JWTBearer())) -> Union[UserRetrieve, Any]:
     """ Get user profile """
+    if payload == None:
+        return generate_unauth_error()
+    
+    msisdn = payload["msisdn"]
+    
     user = db.query(User).filter(User.msisdn==msisdn).first()
+    if not user.refresh_token:
+         return generate_unauth_error()
+     
     return UserRetrieve(
         status="success",
         id=user.id,
@@ -68,9 +83,17 @@ async def get_profile(msisdn = Depends(JWTBearer())) -> UserRetrieve:
         profile_pic_url = user.profile_pic_url)
 
 @router.patch('/profile')
-async def update_profile(user_profile: UserUpdate, msisdn = Depends(JWTBearer())):
+async def update_profile(user_profile: Union[UserUpdate, Any], payload = Depends(JWTBearer())):
     """ Update user profile """
+    if payload == None:
+        return generate_unauth_error()
+    
+    msisdn = payload["msisdn"]
+    
     user = db.query(User).filter(User.msisdn==msisdn).first()
+    if not user.refresh_token:
+         return generate_unauth_error()
+     
     if user_profile.name:
         user.name = user_profile.name
     if user_profile.password:
@@ -81,6 +104,7 @@ async def update_profile(user_profile: UserUpdate, msisdn = Depends(JWTBearer())
         user.profile_pic_url = user_profile.profile_pic_url
     db.commit()
     db.refresh(user)
+    
     return {"status":"success"}
 
 @router.post('/login')
@@ -89,25 +113,33 @@ async def login(msisdn: str = Body(...), password: str = Body(...)) -> dict:
     user = db.query(User).filter(User.msisdn==msisdn).first()
     if user and  password == user.password:
         token = sign_jwt({"msisdn": msisdn})
-        user.refresh_token = token["token"]
+        access_token = token["access_token"]
+        user.refresh_token = token["refresh_token"]
         db.commit()
-        return {"status": "success", "refresh_token": token, "access_token": token}
+        
+        return {"status": "success", "refresh_token": user.refresh_token, "access_token": access_token}
+    
     return Error().dict()
 
 @router.post('/logout')
-async def logtout(msisdn = Depends(JWTBearer())):
+async def logout(payload = Depends(JWTBearer())):
     """ Logout (aka delete refresh token) """
+    if payload == None:
+        return generate_unauth_error()
+    
+    msisdn = payload["msisdn"]
+    
     user = db.query(User).filter(User.msisdn==msisdn).first()
     if user and user.refresh_token:
-        user.refresh_token = ""
+        user.refresh_token = None
         db.commit()
+        
     return {"status":"success"}
 
 
 @router.post('/token')
-async def gen_access_token(refresh_token : Optional[str] = Header(None)):
+async def gen_access_token(refresh_token : Optional[str] = Header(default=None)):
     """ Generate access token from provided refresh token """
-    
     if refresh_token is not None:
         data = decode_jwt(refresh_token)
         if "msisdn" in data:
@@ -115,14 +147,20 @@ async def gen_access_token(refresh_token : Optional[str] = Header(None)):
             user = db.query(User).filter(User.msisdn=="msisdn").first()
             if user is not None:
                 return {"status": "success", "access_token" : sign_jwt({"msisdn": msisdn})}
+            
     return {"status": "failed", "code": 99}
 
 
 @router.post('/delete')
-async def delete(msisdn = Depends(JWTBearer())):
+async def delete(payload = Depends(JWTBearer())):
     """ Delete user """
+    if payload == None:
+        return generate_unauth_error()
+    
+    msisdn = payload["msisdn"]
     user = db.query(User).filter(User.msisdn==msisdn).first()
     assert user
     db.delete(user)
     db.commit()
+    
     return {"status": "success"}
